requirements.txt
pip
py-cord
python-dotenv
datetime